fix: fixed mistakes in tasks
